Move LaneSwitch to own package and separate item
