--- src/components/call/LaneSwitch.scss
@@ -1,46 +1,0 @@
-.LaneSwitch {
-    position: absolute;
-    top: 20px;
-    left: 10px;
-    z-index: 200;
-
-    .LaneSwitch-openLogButton {
-        float: left;
-        width: 40px;
-        height: 40px;
-        background-color: red;
-        border-radius: 50%;
-        box-shadow: 2px 2px 10px rgba(0,0,0,0.3);
-    }
-
-    .LaneSwitch-callList {
-        margin: 0;
-        padding: 0;
-        float: left;
-    }
-
-    .LaneSwitch-callItem {
-        list-style-type: none;
-        color: white;
-        font-weight: bold;
-        width: 40px;
-        height: 40px;
-        border-radius: 50%;
-        margin: 0 0 0 10px;
-        float: left;
-        cursor: pointer;
-        box-shadow: 2px 2px 10px rgba(0,0,0,0.3);
-
-        .Avatar {
-            width: 100%;
-            height: 100%;
-        }
-
-        &:hover {
-            position: relative;
-            left: -1px;
-            top: -1px;
-            box-shadow: 4px 4px 10px rgba(0,0,0,0.4);
-        }
-    }
-}
