+++ src/components/misc/laneSwitch/LaneSwitchItem.scss
@@ -1,0 +1,24 @@
+.LaneSwitchItem {
+    list-style-type: none;
+    color: white;
+    font-weight: bold;
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    margin: 0 0 0 10px;
+    float: left;
+    cursor: pointer;
+    box-shadow: 2px 2px 10px rgba(0,0,0,0.3);
+
+    .Avatar {
+        width: 100%;
+        height: 100%;
+    }
+
+    &:hover {
+        position: relative;
+        left: -1px;
+        top: -1px;
+        box-shadow: 4px 4px 10px rgba(0,0,0,0.4);
+    }
+}
